Handle different Git behaviours with cherry-picks and hooks
This handles errors related to a new Git behaviour which triggers
post-commit hooks during cherry-picks (e.g., during interactive
rebases). With this behaviour it is possible that git-privacy redate is
called while a cherry-pick is still in progress. Consequently, the amend
rewriter fails to perform amends and an error occurs.

This workaround tries to detect ongoing cherry-picks and to await their
completion before triggering the amend. However, some cherry-picks will
inexplicably not conclude while the post-commit hook is not finished.
Hence, this workaround will abort redating after same timeout and
provide a warning to the user.